pyproject.toml
[project]
name = "webchat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.122.0",
    "pydantic-settings>=2.12.0",
    "redis>=7.1.0",
    "uvicorn>=0.38.0",
    "websockets>=15.0.1",
]

[tool.basedpyright]
typeCheckingMode="standard"
